Section3-CloudBuild-CICD/cloudrun-app-cloudbuild/application-files/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
Werkzeug==2.1.2
gunicorn==20.1.0
pandas==2.3.3
pyarrow==22.0.0
pytest==8.4.2
